phase-5/backend/requirements.txt
# ============================================================================
# Phase-3 Backend Dependencies
# ============================================================================

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary==2.9.11
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6

# Validation
pydantic==2.12.5
email-validator>=2.0.0

# Environment Variables
python-dotenv==1.0.0
pydantic-settings==2.7.0

# Phase 3: Real-time & Advanced Features
redis==5.0.1
alembic==1.13.1
google-generativeai==0.8.3
# Phase 5: AI Integration & Advanced Analytics
openai==1.61.0
pgvector==0.3.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
